requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
configparser==3.7.1
Django==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
google-auth==1.6.2
idna==2.8
mysqlclient==1.4.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
